Refactor parameter dictionary helpers into helpers
