+++ roicat/util.py
@@ -250,84 +250,6 @@ def get_default_parameters(
 
     return out
 
-def fill_in_params(
-    d: Dict, 
-    defaults: Dict,
-    verbose: bool = True,
-    hierarchy: List[str] = ['params'], 
-):
-    """
-    In-place. Fills in dictionary ``d`` with values from ``defaults`` if they
-    are missing. Works hierachically.
-    RH 2023
-
-    Args:
-        d (Dict):
-            Dictionary to fill in.
-            In-place.
-        defaults (Dict):
-            Dictionary of defaults.
-        verbose (bool):
-            Whether to print messages.
-        hierarchy (List[str]):
-            Used internally for recursion.
-            Hierarchy of keys to d.
-    """
-    from copy import deepcopy
-    for key in defaults:
-        if key not in d:
-            print(f"Parameter '{key}' not found in params dictionary: {' > '.join([f'{str(h)}' for h in hierarchy])}. Using default value: {defaults[key]}") if verbose else None
-            d.update({key: deepcopy(defaults[key])})
-        elif isinstance(defaults[key], dict):
-            assert isinstance(d[key], dict), f"Parameter '{key}' must be a dictionary."
-            fill_in_params(d[key], defaults[key], hierarchy=hierarchy+[key])
-            
-def check_keys_subset(d, default_dict, hierarchy=['defaults']):
-    """
-    Checks that the keys in d are all in default_dict. Raises an error if not.
-    RH 2023
-
-    Args:
-        d (Dict):
-            Dictionary to check.
-        default_dict (Dict):
-            Dictionary containing the keys to check against.
-        hierarchy (List[str]):
-            Used internally for recursion.
-            Hierarchy of keys to d.
-    """
-    default_keys = list(default_dict.keys())
-    for key in d.keys():
-        assert key in default_keys, f"Parameter '{key}' not found in defaults dictionary: {' > '.join([f'{str(h)}' for h in hierarchy])}."
-        if isinstance(default_dict[key], dict) and isinstance(d[key], dict):
-            check_keys_subset(d[key], default_dict[key], hierarchy=hierarchy+[key])
-
-def prepare_params(params, defaults, verbose=True):
-    """
-    Does the following:
-        * Checks that all keys in ``params`` are in ``defaults``.
-        * Fills in any missing keys in ``params`` with values from ``defaults``.
-        * Returns a deepcopy of the filled-in ``params``.
-
-    Args:
-        params (Dict):
-            Dictionary of parameters.
-        defaults (Dict):
-            Dictionary of defaults.
-        verbose (bool):
-            Whether to print messages.
-    """
-    from copy import deepcopy
-    ## Check inputs
-    assert isinstance(params, dict), f"p must be a dict. Got {type(params)} instead."
-    ## Make sure all the keys in p are valid
-    check_keys_subset(params, defaults)
-    ## Fill in any missing keys with defaults
-    params_out = deepcopy(params)
-    fill_in_params(params_out, defaults, verbose=verbose)
-
-    return params_out
-
 
 def system_info(verbose: bool = False,) -> Dict:
     """
